IOTEX-275 Move network-wide consensus config to genesis config and refactor code (#571)

+++ consensus/scheme/rolldpos/epochctx_test.go
@@ -19,8 +19,8 @@ import (
 
 func TestNewEpochCtx(t *testing.T) {
 	require := require.New(t)
-	numDelegates := uint(24)
-	numSubEpochs := uint(360)
+	numDelegates := uint64(24)
+	numSubEpochs := uint64(360)
 	candidates := []*state.Candidate{}
 	addrs := []string{}
 	for i := 0; i < 24; i++ {
